Refactor to write canon state tag with a single function
Use one function to write to canon, instead of 3. The original idea was
to have the state tag live longer, possibly as a replacement for the
interpreter's internal state, but this didn't really pan out. Instead,
we create them on demand, so it might as well be a single function call.

Signed-off-by: Robert W. Ellenberg <rwe24g@gmail.com>